# services/scenario-manager/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Scenario Manager API",
    "description": "API for managing scenarios, strategies, and investments.",
    "version": "v1"
  },
  "paths": {
...
    "/api/scenarios/{scenarioId}/invest-events": {
      "post": {
        "tags": [
          "InvestEvents"
        ],
        "parameters": [
          {
            "name": "scenarioId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvestEventRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvestEventRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvestEventRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventSeriesResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventSeriesResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventSeriesResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "InvestEvents"
        ],
        "parameters": [
          {
            "name": "scenarioId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventSeriesResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventSeriesResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventSeriesResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/scenarios/{scenarioId}/invest-events/{id}": {
      "get": {
        "tags": [
          "InvestEvents"
        ],
        "parameters": [
          {
            "name": "scenarioId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventSeriesResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventSeriesResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventSeriesResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "InvestEvents"
        ],
        "parameters": [
          {
            "name": "scenarioId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvestEventRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvestEventRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvestEventRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventSeriesResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventSeriesResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventSeriesResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "InvestEvents"
        ],
        "parameters": [
          {
            "name": "scenarioId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/scenarios/{scenarioId}/rebalance-events": {
      "post": {
        "tags": [
          "RebalanceEvents"
        ],
        "parameters": [
          {
            "name": "scenarioId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRebalanceEventRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRebalanceEventRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRebalanceEventRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventSeriesResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventSeriesResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventSeriesResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "RebalanceEvents"
        ],
        "parameters": [
          {
            "name": "scenarioId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventSeriesResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventSeriesResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventSeriesResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/scenarios/{scenarioId}/rebalance-events/{id}": {
      "get": {
        "tags": [
          "RebalanceEvents"
        ],
        "parameters": [
          {
            "name": "scenarioId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventSeriesResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventSeriesResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventSeriesResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "RebalanceEvents"
        ],
        "parameters": [
          {
            "name": "scenarioId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRebalanceEventRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRebalanceEventRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRebalanceEventRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventSeriesResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventSeriesResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventSeriesResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "RebalanceEvents"
        ],
        "parameters": [
          {
            "name": "scenarioId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountTaxStatus": {
        "enum": [
          "NonRetirement",
          "PreTaxRetirement",
          "AfterTaxRetirement"
        ],
        "type": "string"
      },
...
      "CreateInvestEventRequest": {
        "required": [
          "assetAllocation",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "maxLength": 1000,
            "type": "string",
            "nullable": true
          },
          "startYear": {
            "$ref": "#/components/schemas/Distribution"
          },
          "duration": {
            "$ref": "#/components/schemas/Distribution"
          },
          "referenceEventSeriesId": {
            "type": "string",
            "nullable": true
          },
          "startTimingType": {
            "maxLength": 50,
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "orderIndex": {
            "type": "integer",
            "format": "int32"
          },
          "assetAllocation": {
            "type": "object",
            "additionalProperties": {
              "type": "number",
              "format": "double"
            }
          },
          "isGlidePath": {
            "type": "boolean"
          },
          "initialAllocation": {
            "type": "object",
            "additionalProperties": {
              "type": "number",
              "format": "double"
            },
            "nullable": true
          },
          "finalAllocation": {
            "type": "object",
            "additionalProperties": {
              "type": "number",
              "format": "double"
            },
            "nullable": true
          },
          "maximumCash": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CreateRebalanceEventRequest": {
        "required": [
          "assetAllocation",
          "name",
          "targetTaxStatus"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "maxLength": 1000,
            "type": "string",
            "nullable": true
          },
          "startYear": {
            "$ref": "#/components/schemas/Distribution"
          },
          "duration": {
            "$ref": "#/components/schemas/Distribution"
          },
          "referenceEventSeriesId": {
            "type": "string",
            "nullable": true
          },
          "startTimingType": {
            "maxLength": 50,
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "orderIndex": {
            "type": "integer",
            "format": "int32"
          },
          "assetAllocation": {
            "type": "object",
            "additionalProperties": {
              "type": "number",
              "format": "double"
            }
          },
          "isGlidePath": {
            "type": "boolean"
          },
          "initialAllocation": {
            "type": "object",
            "additionalProperties": {
              "type": "number",
              "format": "double"
            },
            "nullable": true
          },
          "finalAllocation": {
            "type": "object",
            "additionalProperties": {
              "type": "number",
              "format": "double"
            },
            "nullable": true
          },
          "targetTaxStatus": {
            "$ref": "#/components/schemas/AccountTaxStatus"
          }
        },
        "additionalProperties": false
      },
...
      "Distribution": {
        "required": [
          "type"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "mean": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "stdev": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "lower": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "upper": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "EventSeriesResponse": {
        "required": [
          "id",
          "name",
          "scenarioId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "scenarioId": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "eventType": {
            "$ref": "#/components/schemas/EventSeriesType"
          },
          "startYear": {
            "$ref": "#/components/schemas/Distribution"
          },
          "duration": {
            "$ref": "#/components/schemas/Distribution"
          },
          "referenceEventSeriesId": {
            "type": "string",
            "nullable": true
          },
          "startTimingType": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "orderIndex": {
            "type": "integer",
            "format": "int32"
          },
          "initialAmount": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "annualChange": {
            "$ref": "#/components/schemas/Distribution"
          },
          "inflationAdjusted": {
            "type": "boolean"
          },
          "userPercentage": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "isSocialSecurity": {
            "type": "boolean"
          },
          "isDiscretionary": {
            "type": "boolean"
          },
          "assetAllocation": {
            "type": "object",
            "additionalProperties": {
              "type": "number",
              "format": "double"
            },
            "nullable": true
          },
          "isGlidePath": {
            "type": "boolean"
          },
          "initialAllocation": {
            "type": "object",
            "additionalProperties": {
              "type": "number",
              "format": "double"
            },
            "nullable": true
          },
          "finalAllocation": {
            "type": "object",
            "additionalProperties": {
              "type": "number",
              "format": "double"
            },
            "nullable": true
          },
          "maximumCash": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "targetTaxStatus": {
            "$ref": "#/components/schemas/AccountTaxStatus"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "EventSeriesType": {
        "enum": [
          "Income",
          "Expense",
          "Invest",
          "Rebalance"
        ],
        "type": "string"
      },
...
    }
  }
}